Add ellipsis to signature of the "updateObject" methods
The "updateObject" methods for SGVariants and SGVariantCounts now use the
same signature as the generic i.e. 'object, ..., verbose=FALSE'.

Note that "updateObject" methods should always have the ellipsis in their
signature. This allows them to be called by other "updateObject" methods
that pass extra arguments in the call. For example the rowRanges() getter
for RangedSummarizedExperiment objects does

    updateObject(x@rowRanges, check=FALSE)

but fails if x@rowRanges is an object that has an "updateObject" method
that doesn't have the ellipsis in its signature.